Backend/requirements.txt
fastapi
python-multipart
httpx
uvicorn
google-genai